Updated user files for consistency

+++ src/main/java/org/example/megahottakes/services/CommentService.java
@@ -1,5 +1,6 @@
 package org.example.megahottakes.services;
 
+import jakarta.transaction.Transactional;
 import org.example.megahottakes.entities.Comment;
 import org.example.megahottakes.entities.HotTake;
 import org.example.megahottakes.entities.User;
@@ -8,6 +9,7 @@ import org.example.megahottakes.repositories.HotTakeRepository;
 import org.example.megahottakes.repositories.UserRepository;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.Set;
 
 @Service
@@ -20,6 +22,7 @@ public class CommentService {
         this.userRepository = userRepository;
         this.commentRepository = commentRepository;
     }
+    @Transactional
     public Comment addComment(Long hotTakeId, Long authorId, String contentOfComment) {
         Comment newComment = new Comment();
         HotTake hotTakeObject = hotTakeRepository.findById(hotTakeId)
@@ -29,13 +32,18 @@ public class CommentService {
         newComment.setAuthor(authorObject);
         newComment.setHotTake(hotTakeObject);
         newComment.setContent(contentOfComment);
+        newComment.setCreatedDate(LocalDateTime.now());
         return commentRepository.save(newComment);
     }
     public Set<Comment> getCommentsByHotTake(Long hotTakeId){
         HotTake hotTake = hotTakeRepository.findById(hotTakeId).orElseThrow(() -> new IllegalArgumentException("The HotTake was not found"));
         return hotTake.getComments();
     }
+    @Transactional
     public void deleteComment(Long commentId){
+        if (!commentRepository.existsById(commentId)){
+            throw new IllegalArgumentException("The Comment was not Found");
+        }
         commentRepository.deleteById(commentId);
     }
 }
